CP-100 Apply new structure rules to settings
