+++ packages/app/src/__deprecated__/renderer/modules/settings/components/app-update-flow/app-update-flow-test-ids.enum.ts
