+++ packages/app/src/__deprecated__/renderer/modules/settings/components/app-update-flow/app-update-flow.container.tsx
@@ -4,9 +4,13 @@
  */
 
 import { connect } from "react-redux"
-import { ReduxRootState, RootState } from "App/__deprecated__/renderer/store"
+import {
+  ReduxRootState,
+  RootState,
+  TmpDispatch,
+} from "App/__deprecated__/renderer/store"
 import { hideModals } from "App/modals-manager/actions"
-import AppForcedUpdateFlow from "App/__deprecated__/renderer/modules/settings/components/app-forced-update-flow.component"
+import AppUpdateFlow from "App/__deprecated__/renderer/modules/settings/components/app-update-flow/app-update-flow.component"
 
 const mapStateToProps = (state: RootState & ReduxRootState) => {
   return {
@@ -15,8 +19,8 @@ const mapStateToProps = (state: RootState & ReduxRootState) => {
   }
 }
 
-const mapDispatchToProps = {
-  hideModals,
-}
+const mapDispatchToProps = (dispatch: TmpDispatch) => ({
+  closeModal: () => dispatch(hideModals()),
+})
 
-export default connect(mapStateToProps, mapDispatchToProps)(AppForcedUpdateFlow)
+export default connect(mapStateToProps, mapDispatchToProps)(AppUpdateFlow)
