+++ packages/app/src/__deprecated__/renderer/modules/settings/components/collecting-data-modal/collecting-data-modal.container.tsx
