+++ packages/app/src/__deprecated__/renderer/modules/settings/components/settings-toggler/settings-toggler-test-ids.enum.ts
