+++ packages/app/src/__deprecated__/renderer/modules/settings/components/settings/settings-ui.component.tsx
@@ -17,8 +17,8 @@ import { noop } from "App/__deprecated__/renderer/utils/noop"
 import styled from "styled-components"
 import { FormattedMessage } from "react-intl"
 import { borderColor } from "App/__deprecated__/renderer/styles/theming/theme-getters"
-import SettingsToggler from "App/__deprecated__/renderer/modules/settings/components/settings-toggler.component"
-import { SettingsTestIds } from "App/__deprecated__/renderer/modules/settings/settings.enum"
+import SettingsToggler from "App/__deprecated__/renderer/modules/settings/components/settings-toggler/settings-toggler.component"
+import { SettingsTestIds } from "App/__deprecated__/renderer/modules/settings/constants/settings.enum"
 import { IconButtonWithPrimaryTooltip } from "App/__deprecated__/renderer/components/core/icon-button-with-tooltip/icon-button-with-primary-tooltip.component"
 import { defineMessages } from "react-intl"
 import { Feature, flags } from "App/feature-flags"
