+++ packages/app/src/__deprecated__/renderer/modules/settings/constants/settings.enum.ts
