+++ packages/app/src/__deprecated__/renderer/modules/settings/helpers/settings-toggler-state.ts
