+++ packages/app/src/modals-manager/containers/modals-manager.container.test.tsx
@@ -9,8 +9,8 @@ import store from "App/__deprecated__/renderer/store"
 import { renderWithThemeAndIntl } from "App/__deprecated__/renderer/utils/render-with-theme-and-intl"
 import ModalsManager from "App/modals-manager/containers/modals-manager.container"
 import { CollectingDataModalTestIds } from "App/__deprecated__/renderer/modules/settings/components/collecting-data-modal/collecting-data-modal-test-ids.enum"
-import { AppForcedUpdateFlowTestIds } from "App/__deprecated__/renderer/modules/settings/components/app-forced-update-flow-test-ids.enum"
-import { AppUpdateFlowTestIds } from "App/__deprecated__/renderer/modules/settings/components/app-update-flow-test-ids.enum"
+import { AppForcedUpdateFlowTestIds } from "App/__deprecated__/renderer/modules/settings/components/app-forced-update-flow/app-forced-update-flow-test-ids.enum"
+import { AppUpdateFlowTestIds } from "App/__deprecated__/renderer/modules/settings/components/app-update-flow/app-update-flow-test-ids.enum"
 
 type Props = ComponentProps<typeof ModalsManager>
